requirements.txt
fastapi==0.95.2
motor==3.1.1
pymongo==4.5.0
uvicorn==0.23.1
pydantic==1.10.12
python-dotenv